Added networking and threads, good times.  The threads do have the issue that they seem to have a really low priority/get smothered by the main thread which is doing all the CPU intensive stuff.  Shuffled a lot of classes around, sorted out logging, fixed a few issues with strings and xml.

+++ library/include/breathe/physics/physics.h
@@ -32,7 +32,7 @@ namespace breathe
 		void AddPhysicsObject(cPhysicsObject *pPhysicsObject);
 		void RemovePhysicsObject(cPhysicsObject *pPhysicsObject);
 		
-		unsigned int GetSize();
+		size_t GetSize();
 
 		typedef std::list<cPhysicsObject * >::iterator iterator;
 		iterator GetObjectListBegin();
